Move FunctionTemplateBrowser to FunctionBrowser directory
