+++ qt/scientific_interfaces/Inelastic/Analysis/FunctionBrowser/ConvTemplateBrowser.h
@@ -6,7 +6,7 @@
 // SPDX - License - Identifier: GPL - 3.0 +
 #pragma once
 
-#include "Analysis/FunctionTemplateBrowser.h"
+#include "Analysis/FunctionBrowser/FunctionTemplateBrowser.h"
 #include "ConvTemplatePresenter.h"
 #include "DllConfig.h"
 #include "FitTypes.h"
