+++ qt/scientific_interfaces/Inelastic/Analysis/FunctionBrowser/IqtTemplateBrowser.h
@@ -6,7 +6,7 @@
 // SPDX - License - Identifier: GPL - 3.0 +
 #pragma once
 
-#include "Analysis/FunctionTemplateBrowser.h"
+#include "Analysis/FunctionBrowser/FunctionTemplateBrowser.h"
 #include "DllConfig.h"
 #include "IqtTemplatePresenter.h"
